weather/requirements.txt
grpcio-tools==1.64.0
fastapi==0.70.0
uvicorn==0.15.0